+++ NEWS.md
@@ -16,6 +16,8 @@
 
 + `is_running()` checks if the current running application is a `{golem}` based application (#366)
 
++ `use_external_file()` allows to add any file to the `www` folder (#295)
+
 ## New features
 
 + `use_external_*()` function don't open files by default (#404)